Reorganize file structure for ProgrammingQuestion and OnlineEditor files
