+++ client/app/bundles/course/assessment/question/programming/components/ExistingDataFile.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
 import {
-  TableHeader, TableHeaderColumn, TableRow, TableRowColumn,
+  TableHeaderColumn, TableRow, TableRowColumn,
 } from 'material-ui/Table';
 import RaisedButton from 'material-ui/RaisedButton';
-import styles from './OnlineEditorView.scss';
 import { grey100, grey300, white } from 'material-ui/styles/colors';
+import styles from './../containers/OnlineEditor/OnlineEditorView.scss';
 import { formatBytes } from './../reducers/utils';
 
 class ExistingDataFile extends React.Component {
@@ -16,16 +16,16 @@ class ExistingDataFile extends React.Component {
     toDelete: PropTypes.bool.isRequired,
     deleteExistingDataFile: PropTypes.func.isRequired,
     isLoading: PropTypes.bool.isRequired,
-    isLast: PropTypes.bool.isRequired
+    isLast: PropTypes.bool.isRequired,
   };
 
   render() {
     // const filename = this.props.data.get('filename');
     // const filesize = this.props.data.get('size');
-    const { filename, filesize, data, toDelete, deleteExistingDataFile, isLast } = this.props;
+    const { filename, filesize, toDelete, deleteExistingDataFile, isLoading, isLast } = this.props;
     const buttonClass = toDelete ? 'fa fa-undo' : 'fa fa-trash';
     const buttonColor = toDelete ? white : grey300;
-    let rowStyle = toDelete ?
+    const rowStyle = toDelete ?
       { textDecoration: 'line-through', backgroundColor: grey100 } : {};
     if (isLast) {
       rowStyle.borderBottom = 'none';
@@ -37,7 +37,7 @@ class ExistingDataFile extends React.Component {
           <RaisedButton
             backgroundColor={buttonColor}
             icon={<i className={buttonClass} />}
-            disabled={this.props.isLoading}
+            disabled={isLoading}
             onClick={() => { deleteExistingDataFile(filename, !toDelete); }}
             style={{ minWidth: '40px', width: '40px' }}
           />
@@ -53,6 +53,6 @@ class ExistingDataFile extends React.Component {
       </TableRow>
     );
   }
-};
+}
 
-export default injectIntl(ExistingDataFile)
+export default injectIntl(ExistingDataFile);
